Try consolidating Pythonic and common to solve time_spec issue
