+++ TIMBER/Framework/include/common.h
@@ -1,7 +1,7 @@
 #ifndef _TIMBER_COMMON
 #define _TIMBER_COMMON
 #ifndef _STRUCT_TIMESPEC
-#define _STRUCT_TIMESPEC 1
+#define _STRUCT_TIMESPEC
 #endif
 
 #include "libarchive/include/archive.h"
@@ -10,6 +10,11 @@
 #include <string>
 #include <iostream>
 #include <sstream>
+#include <algorithm>
+#include <iterator>
+#include <vector>
+#include <stdexcept>
+#include <sys/stat.h>
 #include <boost/filesystem.hpp>
 
 #include <cmath>
@@ -126,6 +131,130 @@ namespace hardware {
      * @return double 
      */
     double invariantMass(RVec<ROOT::Math::PtEtaPhiMVector> vects);
+}
+
+namespace Pythonic {
+    /**
+     * @brief Python-like range function.
+     * https://stackoverflow.com/questions/13152252/is-there-a-compact-equivalent-to-python-range-in-c-stl
+     * 
+     * @tparam IntType 
+     * @param start Starting point, inclusive.
+     * @param stop Stopping point, exclusive.
+     * @param step Defaults to 1.
+     * @return std::vector<IntType> Vector of range with provided step.
+     */
+    template <typename IntType>
+    std::vector<IntType> Range(IntType start, IntType stop, IntType step) {
+        if (step == IntType(0)) {
+            throw std::invalid_argument("step for range must be non-zero");
+        }
+
+        std::vector<IntType> result;
+        IntType i = start;
+        while ((step > 0) ? (i < stop) : (i > stop)) {
+            result.push_back(i);
+            i += step;
+        }
+
+        return result;
+    }
+
+    /**
+     * @brief Python-like range function with step 1.
+     * https://stackoverflow.com/questions/13152252/is-there-a-compact-equivalent-to-python-range-in-c-stl
+     * 
+     * @tparam IntType 
+     * @param start Starting point, inclusive.
+     * @param stop Stopping point, exclusive.
+     * @return std::vector<IntType> Vector of range with step of 1.
+     */
+    template <typename IntType>
+    std::vector<IntType> Range(IntType start, IntType stop) {
+        return Range(start, stop, IntType(1));
+    }
+
+    /**
+     * @brief Python-like range function with step 1 and start assumed to be 0.
+     * https://stackoverflow.com/questions/13152252/is-there-a-compact-equivalent-to-python-range-in-c-stl
+     * 
+     * @tparam IntType 
+     * @param start Starting point, inclusive.
+     * @return std::vector<IntType> Vector of range with step of 1 and start of 0.
+     */
+    template <typename IntType>
+    std::vector<IntType> Range(IntType stop) {
+        return Range(IntType(0), stop, IntType(1));
+    }
+
+    /**
+     * @brief Python-like string splitter based on a delimiter.
+     * 
+     * Adapted from http://www.martinbroadhurst.com/how-to-split-a-string-in-c.html
+     * 
+     * @param str String to split.
+     * @param delim Char to split around.
+     * @return std::vector<std::string> Vector of pieces split around delimiter.
+     */
+    std::vector<std::string> Split(const std::string& str, char delim = ' ');
+
+    // Personal
+    /**
+     * @brief Checks for object in a list.
+     * 
+     * @tparam T 
+     * @param obj Object to look for.
+     * @param list List to look in.
+     * @return bool True or false based on whether object is found in list.
+     */
+    template<typename T>
+    bool InList(T obj, std::vector<T> list) {
+        bool out;
+        auto pos = std::find(std::begin(list), std::end(list), obj);
+        if (pos != std::end(list)){
+            out = true;
+        } else {out = false;}
+        return out;
+    }
+
+    /**
+     * @brief Check for string in another string.
+     * 
+     * @param sub Substring to look for.
+     * @param main String to look in.
+     * @return out  True or false based on whether sub is found in main.
+     * @return false 
+     */
+    bool InString(std::string sub, std::string main);
+
+    /**
+     * @brief Extend vector with another vector (modifies base in-place).
+     * 
+     * @tparam T 
+     * @param base Modif
+     * @param extension 
+     */
+    template<typename T>
+    void Extend(std::vector<T> base, std::vector<T> extension) {
+        for (int i = 0; i < extension.size(); i++) {
+            base.push_back(extension.at(i));
+        }
+    }
+
+    /**
+     * @brief Checks if directory exists.
+     * From https://stackoverflow.com/questions/3828192/checking-if-a-directory-exists-in-unix-system-call
+     * 
+     * @param dirname 
+     * @return exists 
+     */
+    bool IsDir(char* dirname);
+    /**
+     * @brief Prints command and executes via std::system.
+     * 
+     * @param cmd 
+     */
+    void Execute(std::string cmd);
 }
 
 /**
